Refactoring redux app structure
